libs/github_tool/pyproject.toml
[project]
name = "github-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygithub>=2.5.0",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
